pyproject.toml
[tool.poetry]
name = "scan-combiner"
version = "0.1.0"
description = ""
authors = ["Farhan Sareshwala <p.git@fsaresh.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^11.1.0"
pypdf2 = "^3.0.1"

python-dotenv = "^1.0.1"
pikepdf = "^9.5.1"
xmltodict = "^0.14.2"
papersize = "^1.6.0"
requests = "^2.32.4"
zeroconf = "^0.147.0"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
scan-combiner = "scan_combiner.main:combine_files_to_pdf"
